file select backgrounds

+++ src/game/save_file.h
@@ -57,7 +57,7 @@ struct MainMenuSaveData {
     // the older the high score is. This is used for tie-breaking when displaying
     // on the high score screen.
     // u32 coinScoreAges[NUM_SAVE_FILES]; - 96 bits
-    u8 menuRNG;
+    u16 menuRNG;
     u8 soundMode: 2;
 
 #ifdef WIDE
